requirements.txt
streamlit==1.44.1
google-api-python-client==2.168.0
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
sqlalchemy==2.0.40